added 4 new field in CandidateCVResponse and updated process tracking feature

+++ BackEnd/recruitment-service/src/main/java/org/example/recruitmentservice/models/entity/CVAnalysis.java
@@ -28,7 +28,7 @@ public class CVAnalysis {
     @Column
     private String positionName;
 
-    private Double score;
+    private Integer score;
 
     @Column(columnDefinition = "TEXT")
     private String feedback;
